Move the Mode struct to Celt.
It's really just the Celt configuration.
Custom Opus modes (which we don't support)
is essentially just a hack that allows the
application to configure Celt in a special
mode.

Lots of features and Silk are disabled when
using a custom mode and another application
need to implement the same custom mode to
be able to understand the bitstream.

This severely limits the application of
custom modes and they are virtually non
existent in the open.