Reduce code complexity.
By moving some functionality into an inner struct for the decoder
and temporarily creating a heap allocated buffer for transitions,
we are able to not need to use the `SampleBuffer` enum.

We should later judge how common transitions are in hybrid mode,
and if we need to provide a path without allocations.